2-multiagents-langgraph/requirements.txt
langchain
langchain-openai
langchain-google
pydantic
python-dotenv
langgraph
wikipedia
langchain-community
langchain_google_genai
langchain-core